chore: add ActionsShellPage.close()
