+++ catalyst_voices/apps/voices/lib/pages/actions/actions_shell_page.dart
@@ -1,6 +1,8 @@
 import 'package:catalyst_voices/pages/actions/actions_page.dart';
-import 'package:catalyst_voices/pages/co_proposers_consent/co_proposers_consent_page.dart';
-import 'package:catalyst_voices/pages/proposal_approval/proposal_approval_page.dart';
+import 'package:catalyst_voices/pages/actions/co_proposers_consent/co_proposers_consent_page.dart';
+import 'package:catalyst_voices/pages/actions/proposal_approval/proposal_approval_page.dart';
+import 'package:catalyst_voices/routes/routing/actions_route.dart';
+import 'package:catalyst_voices/routes/routing/transitions/transitions.dart';
 import 'package:catalyst_voices/widgets/drawer/voices_drawer.dart';
 import 'package:flutter/material.dart';
 
@@ -26,5 +28,12 @@ class ActionsShellPage extends StatelessWidget {
         child: child,
       ),
     );
+  }
+
+  /// This page is designed to work "as a Drawer". It works in assumption that
+  /// it's route([ActionsShellRoute]) is using [EndDrawerShellPageTransitionMixin] which
+  /// puts child's in [Scaffold] end drawer property.
+  static void close(BuildContext context) {
+    Scaffold.of(context).closeEndDrawer();
   }
 }
